Add lightning color option and height preference
Lightning will strike the highest block in the chunk it is spawned in, randomly choosing one of the highest blocks if multiple exist. Forwent keeping track of chunk data. It required significantly more complex data storage, manipulation, and maintenance that simply searching a chunk at the time of the lightning bolt strike. While the search distance may be expanded to multiple chunks in the future, it is O(n) time to find the highest blocks, where each chunk is 256 operations. For n chunks, it will take O(256n) time, or just O(n) in asymptotic time. Basically, little reason to spend the constant resources managing chunk storage.

Also, the lightning bolt renderer will render the bolt a custom color as specified in the config.

There is still a lot to do in terms of getting to the level of features available in the old version.