Fix lightning color saving and fire spawning
The config file will now store the color of the lightning bolt. It should persist between games now.

Added an option in the config to disable lightning fire. Prevented fire from spawning fire at the original lightning strike position. Needs to be tested for how well it works with fire tick turned on and off, as well as with spawning lightning as only an effect.

Lastly, incremented version strings to the next release version.